bloop

+++ utils/importers/fb3.py
@@ -11,22 +11,28 @@ from bs4 import BeautifulSoup
 import requests
 
 headers = {
-    'cookie': 'fr=103AEBsct6KCQRsmx.AWXg0HaUWnFqQ_3rnAzNqNqao0Y.BdjKWd.Z_.F29.0.0.Be0NDc.AWURYLd1; datr=naWMXY6J_4tOd8nfDzUs7oCI; sb=pKWMXWmDVFUG9iVOfr8b_QM9; c_user=632418911; xs=42%3Af_gMqEFxvrASow%3A2%3A1569498532%3A17543%3A8112; wd=1920x938; presence=EDvF3EtimeF1590743260EuserFA2632418911A2EstateFDsb2F1590725822102EatF1590727055434Et3F_5bDiFA2thread_3a1474557102572072A2EoF1EfF1CAcDiFA2thread_3a537742362963432A2EoF2EfF2CAcDiFA2user_3a623410364A2EoF3EfF3C_5dEutc3F1590716747909G590727055446CEchF_7bCC; act=1590707590692%2F19; spin=r.1002174684_b.trunk_t.1590687066_s.1_v.2_',
     'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:76.0) Gecko/20100101 Firefox/76.0'
 } # paste the cookie here
 cache_folder = Path("D:\\temp\\cache")
+save_file = Path("d:\\temp\\out.json")
 last_html = Path("D:\\temp\\cache\\last.html")
 base_url = "https://m.facebook.com/profile/timeline/stream/?profile_id=632418911"
-base_url = "https://m.facebook.com/profile/timeline/stream/?cursor=AQHR8yQCpFyzapKSGPyVN4L49pjzfdK8CIGayXesbXzLJpkqIuytg3cWxo1NqjAa3fPH8fQdTfYuzCMo8ORHQzpTbTMw_AOwwQ4NkMGraHyfBjr9a3Bb-U6AwQb0R3WAyQaL&profile_id=632418911&replace_id=u_0_0"
+# base_url = "https://m.facebook.com/profile/timeline/stream/?cursor=AQHRZZP0w6Ci0lRZKPNLsNpupJHCFFX9QpP9h4J68DErLxTQZXFKsepZ3CPxaj0Lb7E4_yWjKJq2kbLviquAPcMAH-RHH3hyfeeW0fXsVBq54uqLPlyXq8n5UIeu_paa7EEj&profile_id=632418911&replace_id=u_0_0"
 FOCUS_ID = ""
 SINGLE = False
-USE_CACHE = True
+USE_CACHE_COMMENTS = True
+USE_CACHE_MAIN = False
 ME = "Roy Tang"
+parser = "lxml"
+
+
+def get_text(soup):
+    return ' '.join(soup.findAll(text=True))
 
 def get_comments_detail(url, story_id, page=0, replyto=""):
 
     out_html = cache_folder / "html" / "detail" / (story_id + "-" + str(page) + replyto + ".html")
-    if out_html.exists() and USE_CACHE:
+    if out_html.exists() and USE_CACHE_COMMENTS:
         with out_html.open(encoding="UTF-8") as f:
             text = f.read()
     else:
@@ -45,7 +51,7 @@ def get_comments_detail(url, story_id, page=0, replyto=""):
             }]
 
     out_comments = []
-    soup = BeautifulSoup(text, "html.parser")
+    soup = BeautifulSoup(text, parser)
     selector = "#m_story_permalink_view > div:nth-child(2) > div > div:nth-child(4) > div"
     if len(replyto) > 0:
         selector = "#root > div > :nth-child(3) > div"
@@ -75,7 +81,7 @@ def get_comments_detail(url, story_id, page=0, replyto=""):
         #         print(a.name, a.text)
         #     except:
         #         print(str(a))
-        comment["text"] = c.contents[0].contents[1].text
+        comment["text"] = get_text(c.contents[0].contents[1])
         # meta row contains the date
         if len(c.contents[0].contents) >= 4:
             for abbr in c.contents[0].contents[3].select("abbr"):
@@ -95,7 +101,7 @@ def parse_container(child):
     item = {"attachments":[]}
 
     story_id = None
-
+    comments_link = None
     # find the permalink via the comments anchor
     for link in child.select("a"):
         if link.text.find("Comment") >= 0:
@@ -111,8 +117,10 @@ def parse_container(child):
     if child.get("data-ft") is not None:
         item["metadata"] = json.loads(child["data-ft"])
         if story_id is None:
-            # TODO: Parse the rest of this page correctly
             story_id = item["metadata"].get("mf_story_key", item["metadata"].get("top_level_post_id"))
+
+    if story_id is None:
+        return []
 
     item["permalink"] = "https://www.facebook.com/stephen.roy.tang/posts/" + story_id
 
@@ -129,17 +137,18 @@ def parse_container(child):
 
     if child.text.find("for your birthday") >= 0:
         item["type"] = "bday"
-        item["details"] = child.text
+        item["details"] = get_text(child)
         return [item]
     children2 = child.contents
     header = None
     body = None
     footer = None
+    # print(child.prettify())
     print(len(children2))
     if len(children2) == 2:
         footer = children2[1]
         body = children2[0]
-        if body.text.find("posted on your") >= 0 or body.text.find("posted from") >= 0 or body.text.find(" is playing "):
+        if body.text.find("posted on your") >= 0 or body.text.find("posted from") >= 0 or body.text.find(" is playing ") >= 0:
             # recursing 
             items = []
             for article in footer.select("article"):
@@ -158,13 +167,26 @@ def parse_container(child):
     elif len(children2) == 1:
         print("ERROR")
         item["type"] = "error"
+        item["desc"] = get_text(child)
         return [item]
 
     date_containers = footer.select("abbr")
     for dc in date_containers:
         item["date"] = dc.text
-    print(body)
+    print(body.text)
     actual_contents = body.contents
+
+    if len(actual_contents) == 1:
+        if header is not None and header.text.find(" is playing "):
+            item["text"] = get_text(body)
+            for link in body.select("a"):
+                item["url"] = link["href"]
+            items = []
+            for article in footer.select("article"):
+                items.extend(parse_container(article))
+            item["subitems"] = items
+            return [item]
+
     item["text"] = actual_contents[1].text
     for link in actual_contents[0].select("a"):
         item["poster_url"] = link["href"].split("?")[0]
@@ -173,7 +195,7 @@ def parse_container(child):
             break
 
     # get comments, but only if it's my content
-    if not comments_link.text.startswith("Comment") and item["poster"] == ME:
+    if comments_link is not None and not comments_link.text.startswith("Comment") and item["poster"] == ME:
         item["comments"] = get_comments_detail(comments_link_url, story_id)
 
 
@@ -185,16 +207,18 @@ def parse_container(child):
             item["repost"] = True
             for link in actual_contents[2].select("h3 a"):
                 poster = link.text
-                poster_url = link["href"].split("?")[0]
-                poster_url = "https://www.facebook.com" + poster_url
+                raw_poster_url = link["href"].split("?")[0]
+                poster_url = "https://www.facebook.com" + raw_poster_url
                 break # only the first link is the poster
             all_details = []
             for link in actual_contents[2].select("a"):
                 link_text = link.text
-                if link_text == poster:
-                    continue # we already got this link
                 url = link.get("href")
                 if url is None:
+                    continue
+                raw_url = url.split("?")[0]
+                # skip any links to the poster profile
+                if link_text == poster or raw_url == raw_poster_url:
                     continue
                 # compose the permalink
                 u = urlparse(url)
@@ -207,12 +231,16 @@ def parse_container(child):
                 if src_url.startswith("https://lm.facebook.com"): # external link
                     u = urlparse(src_url)
                     qparams = parse_qs(u.query)
-                    details = qparams.get("u")
+                    details = {
+                        "url": qparams.get("u")
+                    }
                 else:
                     src_url = link["href"]
                     if not src_url.startswith("http"):
                         src_url = "https://m.facebook.com" + src_url
                     details = get_link_details(src_url)
+                details["text"] = get_text(link)
+                details["src_url"] = src_url
 
                 all_details.append(details)
 
@@ -220,9 +248,8 @@ def parse_container(child):
                 "type": "repost",
                 "poster": poster,
                 "poster_url": poster_url,
-                "src_url": src_url,
                 "url": url,
-                "desc": actual_contents[2].text,
+                "desc": get_text(actual_contents[2]),
                 "details": all_details
             })
         else:
@@ -281,7 +308,7 @@ def parse_container(child):
                     "url": "",
                     "img_url": img_url,
                     "downloaded": str(download_to), 
-                    "desc": actual_contents[2].text
+                    "desc": get_text(actual_contents[2])
                 })
     return [item]
 
@@ -295,31 +322,35 @@ def get_timeline():
             index = index + 1
             out_html = cache_folder / "html" / (str(index) + ".html")
 
-            req = urllib.request.Request(url_next, headers=headers)
-            next_link = ""
-            with urllib.request.urlopen(req) as url:
-                text = url.read()
-                with last_html.open("w", encoding="UTF-8") as f:
-                    f.write(text.decode('utf-8'))
-                with out_html.open("w", encoding="UTF-8") as f:
-                    f.write("<!-- " + url_next + " //-->")
-                    f.write(text.decode('utf-8'))
-                soup = BeautifulSoup(text, "html.parser")
-                # print(soup)
-                # break
-                links = soup.find_all("a")
-                for link in links:
-                    if link.text.find("See More Stories") >= 0:
-                        next_link = "https://m.facebook.com" + link["href"]
-                containers = soup.select("#structured_composer_async_container > *:nth-child(1)")
-                for c in containers:
-                    for child in c.contents:
-                        # print(json.dumps(item, indent=2))
-                        for item in parse_container(child):
-                            if item is not None:
-                                item["source_url"] = url_next
-                                item["source_downloaded"] = str(out_html)
-                            items.append(item)
+            if out_html.exists() and USE_CACHE_MAIN:
+                with out_html.open(encoding="UTF-8") as f:
+                    text = f.read()
+            else:
+                req = urllib.request.Request(url_next, headers=headers)
+                next_link = ""
+                with urllib.request.urlopen(req) as url:
+                    text = url.read()
+                    with last_html.open("w", encoding="UTF-8") as f:
+                        f.write(text.decode('utf-8'))
+                    with out_html.open("w", encoding="UTF-8") as f:
+                        f.write("<!-- " + url_next + " //-->")
+                        f.write(text.decode('utf-8'))
+            soup = BeautifulSoup(text, parser)
+            # print(soup)
+            # break
+            links = soup.find_all("a")
+            for link in links:
+                if link.text.find("See More Stories") >= 0:
+                    next_link = "https://m.facebook.com" + link["href"]
+            containers = soup.select("#structured_composer_async_container > *:nth-child(1)")
+            for c in containers:
+                for child in c.contents:
+                    # print(json.dumps(item, indent=2))
+                    for item in parse_container(child):
+                        if item is not None:
+                            item["source_url"] = url_next
+                            item["source_downloaded"] = str(out_html)
+                        items.append(item)
             url_next = next_link
             print("#### " + url_next)
             if SINGLE:
@@ -327,7 +358,7 @@ def get_timeline():
     # except Exception as e:
     #     print(e)
     finally:
-        with Path("d:\\temp\\out.json").open("w", encoding="UTF-8") as f:
+        with save_file.open("w", encoding="UTF-8") as f:
             f.write(json.dumps(items, indent=2))
         print("Done")
 
@@ -347,18 +378,19 @@ def get_link_details(url):
     try: 
         with urllib.request.urlopen(req) as urlr:
             text = urlr.read()
-            soup = BeautifulSoup(text, "html.parser")
+            soup = BeautifulSoup(text, parser)
             # print(soup.prettify())
             for i in soup.select("#MPhotoActionbar"):
                 # indicates a photo
                 mode = "PHOTO"
+                desc = ""
                 for d in soup.select("div.msg"):
                     el = d.contents[2]
                     # sometimes there's a space that gets returned as d.contents[2], 
                     # when that happens skip to next sibling
                     while el.name is None:
                         el = el.next_sibling
-                    desc = el.text
+                    desc = get_text(el)
                 imgs = soup.select("img.img")
                 out_url = imgs[1]["src"]
 
@@ -380,6 +412,7 @@ def get_link_details(url):
                     "type": "photo",
                     "desc": desc,
                     "url": out_url,
+                    "orig_url": url,
                     "downloaded": str(download_to)
                 }
 
@@ -391,8 +424,16 @@ def get_link_details(url):
                 }
 
             # print(soup.prettify())
+
+            if soup.text.find("You can try again later.") >= 0:
+                return {
+                    "type": "rate-limited",
+                    "url": url
+                }
+
             return {
                 "type": "ERROR",
+                "desc": get_text(soup),
                 "url": url
             }
     except urllib.error.HTTPError as e:
@@ -402,8 +443,23 @@ def get_link_details(url):
             "url": url
         }
 
+def verify():
+    with save_file.open(encoding="UTF-8") as f:
+        data = json.loads(f.read())
 
-get_timeline()
+        print(len(data))
 
-# src_url = "https://m.facebook.com/story.php?story_fbid=10158244363002497&id=508217496&_ft_=mf_story_key.10158473723063912%3Atop_level_post_id.10158473723063912%3Atl_objid.10158473723063912%3Acontent_owner_id_new.632418911%3Aoriginal_content_id.10158244363002497%3Aoriginal_content_owner_id.508217496%3Athrowback_story_fbid.10158473723063912%3Astory_location.4%3Athid.632418911%3A306061129499414%3A2%3A0%3A1590994799%3A3441436630452290644&__tn__=%2AsH-R"
-# print(get_link_details(src_url))
+        posters = {}
+        for item in data:
+            poster = item.get("poster", "")
+            if poster not in posters:
+                posters[poster] = 1
+            else:
+                posters[poster] = posters[poster] + 1
+
+        print(json.dumps(posters, indent=2))
+
+verify()
+
+# get_timeline()
+
